+++ src/common/src/ConfirmDialog/ConfirmDialog.test.tsx
@@ -1,54 +1,81 @@
 import React from 'react';
-import { render } from '../../../test-utils/customRender';
-import ConfirmDialog, { IState, confirm } from './ConfirmDialog';
-import userEvent from '@testing-library/user-event';
+import { screen } from '@testing-library/dom';
+import { fireEvent, render } from '../../../test-utils/customRender';
+import ConfirmDialog, { confirm, IState, Props } from './ConfirmDialog';
 
-describe('<ConfirmDialog/>', () => {
-	let getByTestId: Function;
-	let getByText: any;
-	const onCancelMock = jest.fn();
-	const onConfirmMock = jest.fn();
+describe('<ConfirmDialog />', () => {
+	const onClick = jest.fn();
 
-	const props = {
-		title: 'TestDialog',
-		message: 'Are you sure you want to delete this file?',
-		textCancelButton: 'Cancel',
-		textConfirmButton: 'Confirm',
-		onCancel: onCancelMock,
-		onConfirm: onConfirmMock,
-	} as IState;
+	const defaultArg = {
+		title: 'Test Title',
+		message: 'Test Message',
+	};
 
-	beforeEach(() => {
-		({ getByTestId, getByText } = render(
-			<div>
-				<button data-testid="open-dialog" onClick={() => confirm(props)} />
-				<ConfirmDialog />
-			</div>,
-		));
-		const button = getByTestId('open-dialog');
-		userEvent.click(button);
+	const callbackMocks = {
+		onCancel: onClick,
+		onConfirm: onClick,
+	};
+
+	const clickAndRenderDialog = (args: Partial<IState> = defaultArg, props: Partial<Props> = {}) => {
+		render(
+			<>
+				<button data-testid="open-dialog" onClick={() => confirm(args as IState)} />
+				<ConfirmDialog {...props} />
+			</>,
+		);
+		const button = screen.getByTestId('open-dialog');
+		fireEvent.click(button);
+	};
+
+	// checks if the dialog renders
+	test.each([
+		['default', 'Test Waarschuwing', 'Weet u zeker dat u dit bestand wilt verwijderen?', 'Ja', 'Nee'],
+		['custom', 'Test Warning', 'Are you sure you want to delete this file?', 'Cancel', 'Confirm'],
+	])('Renders %s dialog without close button', (testCase, title, message, cancelLabel, confirmLabel) => {
+		clickAndRenderDialog(
+			{
+				title: title,
+				message: message,
+				textCancelButton: cancelLabel,
+				textConfirmButton: confirmLabel,
+				...callbackMocks,
+			},
+			{},
+		);
+		expect(screen.getByText(title)).toBeInTheDocument();
+		expect(screen.getByText(message)).toBeInTheDocument();
+		expect(screen.getByText(cancelLabel)).toBeInTheDocument();
+		expect(screen.getByText(confirmLabel)).toBeInTheDocument();
+		expect(screen.queryByTestId('modal-close-button')).not.toBeInTheDocument();
 	});
 
-	afterEach(() => {
-		jest.resetAllMocks();
+	// checks if all buttons are clickable with passed methods
+	test.each([
+		['confirm', 'confirm-button'],
+		['cancel', 'cancel-button'],
+		['close', 'modal-close-button'],
+	])('Test %s button', (testCase, testId) => {
+		clickAndRenderDialog(
+			{
+				...defaultArg,
+				...callbackMocks,
+			},
+			{
+				hideCloseButton: false,
+			},
+		);
+		fireEvent.click(screen.getByTestId(testId));
+		expect(onClick).toHaveBeenCalled();
 	});
 
-	it('should render the ConfirmDialog correctly', () => {
-		expect(getByText('TestDialog')).toBeInTheDocument();
-		expect(getByText('Are you sure you want to delete this file?')).toBeInTheDocument();
-		expect(getByText('Cancel')).toBeInTheDocument();
-		expect(getByText('Confirm')).toBeInTheDocument();
+	// checks if a dialog without a message renders
+	test('Dialog should not render', () => {
+		clickAndRenderDialog(callbackMocks);
+		expect(screen.queryByTestId('confirm-dialog')).toBeNull();
 	});
 
-	it('should call the onCancel function when clicking the cancelButton', () => {
-		const cancelButton = getByTestId('cancel-button');
-		userEvent.click(cancelButton);
-		expect(onCancelMock).toBeCalled();
-	});
-
-	it('should call the onConfirm function when clicking the confirmButton', () => {
-		const confirmButton = getByTestId('confirm-button');
-		userEvent.click(confirmButton);
-		expect(onConfirmMock).toBeCalled();
+	test('Should not show close button', () => {
+		clickAndRenderDialog();
+		expect(screen.queryByTestId('modal-close-button')).not.toBeInTheDocument();
 	});
 });